Date_python/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pymysql==1.1.0
python-dotenv==1.0.1
sqlalchemy==2.0.29
pydantic==2.6.4
requests==2.31.0
passlib==1.7.4
bcrypt==4.1.2
pyjwt==2.8.0